seperate useUserProfile
