+++ src/hooks/api/auth/useGetUserInfo.ts
@@ -1,31 +1,17 @@
 import { fetchGetUserInfo } from '@/apis/auth.api';
+import { IUserProfile } from '@/models/user.model';
 import useAuthStore from '@/store/authStore';
-import useUserStore from '@/store/userStore';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { IUserInfo } from '@/models/user.model';
+import { useQuery } from '@tanstack/react-query';
 
 const useGetUserProfile = () => {
-  const queryClient = useQueryClient();
   const { isLoggedIn } = useAuthStore();
-  const { setProfile } = useUserStore();
-  const { data } = useQuery({
+  const { data } = useQuery<IUserProfile>({
     queryKey: ['user', 'profile'],
     queryFn: fetchGetUserInfo,
     enabled: isLoggedIn,
   });
 
-  const upDateUserPoint = (point: number) => {
-    const queryKey = ['user', 'profile'];
-    const userProfile: IUserInfo | undefined =
-      queryClient.getQueryData(queryKey);
-    if (userProfile) {
-      const shallow: IUserInfo = { ...userProfile };
-      shallow.point += point;
-      queryClient.setQueryData(queryKey, shallow);
-    }
-  };
-
-  return { userProfile: data, upDateUserPoint };
+  return { userProfile: data };
 };
 
 export default useGetUserProfile;
