+++ src/models/user.model.ts
@@ -1,4 +1,4 @@
-export interface IUserInfo {
+export interface IUserProfile {
   name: string;
   email: string;
   point: number;
